# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel REST API V4 - Master-Detail Architecture",
        "description": "REST API untuk Angular Frontend dengan Master-Detail Architecture (Many-to-Many)<br><br>\n *                  <strong>Fitur V4:</strong><br>\n *                  ✅ Business Units & Menus sebagai Master Data<br>\n *                  ✅ Many-to-Many relationships (User ↔ BU, User ↔ Menu)<br>\n *                  ✅ Dropdown API untuk manage user access<br>\n *                  ✅ Dynamic menu & BU assignment<br><br>\n *                  <strong>Test Accounts:</strong><br>\n *                  - admin / Admin123 (Full access)<br>\n *                  - user1 / User123 (Batam, Jakarta + 3 menus)<br>\n *                  - user2 / User123 (Surabaya + 2 menus)",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/customers/bulk-delete": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Bulk delete customers",
                "description": "Delete multiple customers at once. All authenticated users can perform this action.",
                "operationId": "0dd433889093e5d3f16d81e371123464",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ids"
                                ],
                                "properties": {
                                    "ids": {
                                        "description": "Array of customer IDs to delete",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bulk delete successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "3 customers deleted successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "deleted_count": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "failed_count": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "deleted_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    },
                                                    "example": [
                                                        1,
                                                        2,
                                                        3
                                                    ]
                                                },
                                                "failed_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    },
                                                    "example": []
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Business unit not selected or unauthorized access"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}